SimpleVideoGUI/requirements.txt
SimpleVideoGUI

pyqt==5.6.0
opencv3==3.1.0
numpy==1.11.2